Clean up GlobalKVStore and Historian code
